veiliger implementatie
git-svn-id: svn+ssh://s4.hekkelman.net/home/maarten/svnroot/salt/trunk/libassh@281 49180b5e-4fd1-4af1-8649-882265d54e9a

+++ assh/packet.hpp
@@ -166,7 +166,7 @@ class ipacket
 
 	void			decompress(compression_helper& decompressor, boost::system::error_code& ec);
 	
-	size_t			size() const						{ return m_length; }
+	uint32			size() const						{ return m_length; }
 	
 	void			append(const std::vector<uint8>& block);
 	std::size_t		read(const char* data, std::size_t size);
